test: remove symlinks from runtime_impl_test test_data. (#626)
Bazel doesn't like symlinks in test data deps (#415). This patch fixes this by removing the symlinks
from the test inputs, copying the test inputs to the writable temporary directory and then creating
the symlinks during test execution.

Using ugly system(3) instead of a shell script wrapper as this is a unit test and we shouldn't need
to wrap it.